Move refactor into admin/users
