+++ src/App.tsx
@@ -4,10 +4,10 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import createQueryClient from "./lib/createQueryClient";
 import { UserProvider } from "./lib/hooks/useUser";
 import HomePage from "./pages";
-import UserAdd from "./pages/admin/AddUser";
 import AdminPage from "./pages/admin/AdminPage";
-import EditUser from "./pages/admin/EditUser";
-import UsersOverview from "./pages/admin/UsersOverview";
+import UserAdd from "./pages/admin/users/AddUser";
+import EditUser from "./pages/admin/users/EditUser";
+import UsersOverview from "./pages/admin/users/UsersOverview";
 import EmployeePage from "./pages/EmployeePage";
 import Login from "./pages/Login";
 import { PageNotFound } from "./pages/PageNotFound";
