+++ src/pages/admin/users/AddUser.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Layout from "../../components/Layout";
-import UserForm from "../../components/UserForm";
-import useAxios from "../../lib/hooks/useAxios";
-import { useAdmin } from "../../lib/hooks/useRole";
+import Layout from "../../../components/Layout";
+import UserForm from "../../../components/UserForm";
+import useAxios from "../../../lib/hooks/useAxios";
+import { useAdmin } from "../../../lib/hooks/useRole";
 
 const UserAddForm = () => {
   const [email, setEmail] = useState("");
